+++ src/storage/mod.rs
@@ -1,19 +1,22 @@
 mod builder;
 mod config;
 mod core;
+mod key;
 mod observer;
 mod observer_worker;
-mod key;
+mod read_result;
 
 pub use self::{
     builder::Builder,
     core::Storage,
-    key::{Key, RefKey, ArrayKey},
+    key::{ArrayKey, Key, RefKey},
     observer::ActiveBlobPred,
     observer::ActiveBlobStat,
+    read_result::{BlobRecordTimestamp, ReadResult},
 };
 
 mod prelude {
+    pub(crate) use async_std::sync::RwLock as ASRwLock;
     pub(crate) use {
         super::{
             config::Config, core::Inner, observer::Msg, observer::Observer,
@@ -22,5 +25,4 @@ mod prelude {
         },
         crate::prelude::*,
     };
-    pub(crate) use async_std::sync::RwLock as ASRwLock;
 }